// types.ts
export enum ClothingCategory {
    Shirt = 'Shirt',
    Kurta = 'Kurta',
    TShirt = 'T-Shirt',
    Dress = 'Dress',
    Jacket = 'Jacket',
    Saree = 'Saree',
}

export enum ModelPose {
    Standing = 'Standing',
    Walking = 'Walking',
    Sitting = 'Sitting',
    Action = 'Dynamic Action',
    HandsOnHips = 'Hands on Hips',
    Leaning = 'Leaning',
    LookingAway = 'Looking Away',
}

export enum BackgroundStyle {
    Studio = 'Studio',
    Outdoor = 'Outdoor',
    Abstract = 'Abstract',
}

export enum ModelGender {
    Female = 'Female',
    Male = 'Male',
    Unspecified = 'Unspecified',
}

export enum ModelAppearance {
    Indian = 'Indian',
}